requirements.txt
pycryptodome~=3.16.0
requests~=2.28.1
